Refactored the importdata functions
