+++ setup.ts
@@ -5,9 +5,6 @@ import fs from "fs";
 import path from "path";
 import * as cryptolib from "crypto";
 import inquirer from "inquirer";
-import mongodb from "mongodb";
-import { exec } from "child_process";
-import type { ExecException } from "child_process";
 /* eslint-disable */
 import {
   askForMongoDBUrl,
@@ -29,6 +26,7 @@ import {
   setImageUploadSize,
   validateImageFileSize,
 } from "./src/setup/setImageUploadSize";
+import { importData } from "./src/setup/importData";
 /* eslint-enable */
 
 dotenv.config();
@@ -486,7 +484,7 @@ export async function recaptchaSiteKey(): Promise<void> {
 /**
  * The `abort` function logs a message and exits the process.
  */
-function abort(): void {
+export function abort(): void {
   console.log("\nSetup process aborted. 🫠");
   process.exit(1);
 }
@@ -535,85 +533,6 @@ export async function twoFactorAuth(): Promise<void> {
     config.MAIL_USERNAME = email;
     config.MAIL_PASSWORD = password;
     updateEnvVariable(config);
-  }
-}
-
-//Checks if the data exists and ask for deletion
-/**
- * The function `shouldWipeExistingData` checks if there is existing data in a MongoDB database and prompts the user to delete
- * it before importing new data.
- * @param url - The `url` parameter is a string that represents the connection URL for the
- * MongoDB database. It is used to establish a connection to the database using the `MongoClient` class
- * from the `mongodb` package.
- * @returns The function returns a Promise<boolean>.
- */
-async function shouldWipeExistingData(url: string): Promise<boolean> {
-  let shouldImport = false;
-  const client = new mongodb.MongoClient(url, {
-    useNewUrlParser: true,
-    useUnifiedTopology: true,
-  });
-  try {
-    await client.connect();
-    const db = client.db();
-    const collections = await db.listCollections().toArray();
-
-    if (collections.length > 0) {
-      const { confirmDelete } = await inquirer.prompt({
-        type: "confirm",
-        name: "confirmDelete",
-        message:
-          "We found data in the database. Do you want to delete the existing data before importing?",
-      });
-
-      if (confirmDelete) {
-        for (const collection of collections) {
-          await db.collection(collection.name).deleteMany({});
-        }
-        console.log("All existing data has been deleted.");
-        shouldImport = true;
-      } else {
-        console.log("Deletion & import operation cancelled.");
-      }
-    } else {
-      shouldImport = true;
-    }
-  } catch (error) {
-    console.error("Could not connect to database to check for data");
-  }
-  client.close();
-  return shouldImport;
-}
-
-//Import sample data
-/**
- * The function `importData` imports sample data into a MongoDB database if the database URL is provided and if it
- * is determined that existing data should be wiped.
- * @returns The function returns a Promise that resolves to `void`.
- */
-async function importData(): Promise<void> {
-  if (!process.env.MONGO_DB_URL) {
-    console.log("Couldn't find mongodb url");
-    return;
-  }
-  const shouldImport = await shouldWipeExistingData(process.env.MONGO_DB_URL);
-
-  if (shouldImport) {
-    console.log("Importing sample data...");
-    await exec(
-      "npm run import:sample-data",
-      (error: ExecException | null, stdout: string, stderr: string) => {
-        if (error) {
-          console.error(`Error: ${error.message}`);
-          abort();
-        }
-        if (stderr) {
-          console.error(`Error: ${stderr}`);
-          abort();
-        }
-        console.log(`Output: ${stdout}`);
-      },
-    );
   }
 }
